refactor: move SearchFilters interface definition to types.ts for better organization
